+++ src/RotationSpline.cc
@@ -62,7 +62,7 @@ Quaterniond RotationSpline::Interpolate(const unsigned int _fromIndex,
 {
   // Bounds check
   if (_fromIndex >= this->dataPtr->points.size())
-    return Quaterniond(IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF);
+    return Quaterniond(INF_D, INF_D, INF_D, INF_D);
 
   if ((_fromIndex + 1) == this->dataPtr->points.size())
   {
@@ -167,7 +167,7 @@ void RotationSpline::RecalcTangents()
 /////////////////////////////////////////////////
 const Quaterniond &RotationSpline::Point(const unsigned int _index) const
 {
-  static Quaterniond inf(IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF);
+  static Quaterniond inf(INF_D, INF_D, INF_D, INF_D);
 
   if (this->dataPtr->points.empty())
     return inf;