+++ src/Spline.cc
@@ -105,7 +105,7 @@ Vector3d Spline::Interpolate(const unsigned int _fromIndex,
 {
   // Bounds check
   if (_fromIndex >= this->dataPtr->points.size())
-    return Vector3d(IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF);
+    return Vector3d(INF_D, INF_D, INF_D);
 
   if ((_fromIndex + 1) == this->dataPtr->points.size())
   {
@@ -231,7 +231,7 @@ void Spline::RecalcTangents()
 Vector3d Spline::Point(const unsigned int _index) const
 {
   if (_index >= this->dataPtr->points.size())
-    return Vector3d(IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF);
+    return Vector3d(INF_D, INF_D, INF_D);
 
   return this->dataPtr->points[_index];
 }
@@ -240,7 +240,7 @@ Vector3d Spline::Point(const unsigned int _index) const
 Vector3d Spline::Tangent(const unsigned int _index) const
 {
   if (_index >= this->dataPtr->tangents.size())
-    return Vector3d(IGN_DBL_INF, IGN_DBL_INF, IGN_DBL_INF);
+    return Vector3d(INF_D, INF_D, INF_D);
 
   return this->dataPtr->tangents[_index];
 }