+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Ignition Math 2.x
 
+1. Added OrientedBox
+    * [Pull request 146](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/146)
+
 1. Added an assignment operator to the Frustum class.
     * [Pull request 144](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/144)