+++ Changelog.md
@@ -1,3 +1,7 @@
+## Ignition Math 4.x
+
+### Ignition Math 4.x.x
+
 ## Ignition Math 3.x
 
 ### Ignition Math 3.x.x